Remove babel polyfills
According to babel/babel#7267 it seems that library authors probably
don't want to include polyfills.

After a bunch of problems with babel and core-js and the now familiar
`Module not found: Error: Can't resolve
'core-js/modules/es7.string.pad-end'` error I'm inclined to agree.
Polyfills need to be the concern of the consuming application.